microservice/requirements.txt
fastapi
uvicorn
python-multipart
pillow
ultralytics
numpy
redis
rq